Seperate SNMP message process module
