+++ core/snmp_msg_in.c
@@ -109,462 +109,6 @@ snmp_datagram_clear(struct snmp_datagram *sdg)
   INIT_LIST_HEAD(&sdg->vb_out_list);
 }
 
-static void
-snmp_response(struct snmp_datagram *sdg)
-{
-  snmpd_send_response(sdg);
-}
-
-static void
-mib_get(struct snmp_datagram *sdg, struct var_bind *vb_in, struct oid_search_res *ret_oid)
-{
-  struct mib_view *view = NULL;
-  struct mib_community *community = NULL;
-  struct mib_user *user = NULL;
-
-  /* Access control */
-  if (sdg->version >= 3) {
-    user = mib_user_search(sdg->user_name);
-    if (user == NULL) {
-      ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-    }
-  } else {
-    community = mib_community_search(sdg->context_name);
-    if (community == NULL) {
-      ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-    }
-  }
-
-  /* Traverse all availble views according to community or user */
-  for (; ;) {
-    if (sdg->version >= 3) {
-        view = mib_user_next_view(user, MIB_ACES_READ, view);
-    } else {
-        view = mib_community_next_view(community, MIB_ACES_READ, view);
-    }
-
-    /* End of mib view */
-    if (view == NULL) {
-      /* Duplicate original oid when result not found */
-      ret_oid->oid = oid_dup(vb_in->oid, vb_in->oid_len);
-      ret_oid->id_len = vb_in->oid_len;
-      return;
-    }
-
-    mib_tree_search(view, vb_in->oid, vb_in->oid_len, ret_oid);
-    if ((!ret_oid->err_stat && MIB_TAG_VALID(tag(&ret_oid->var))) || oid_cmp(vb_in->oid, vb_in->oid_len, view->oid, view->id_len) < 0) {
-      /* Gotcha or given oid ahead of all views */
-      return;
-    }
-    /* Free temporary result */
-    free(ret_oid->oid);
-  }
-}
-
-/* GET request function */
-static void
-snmp_get(struct snmp_datagram *sdg)
-{
-  struct list_head *curr, *next;
-  struct var_bind *vb_in, *vb_out;
-  struct oid_search_res ret_oid;
-  uint32_t oid_len, len_len, val_len;
-  uint32_t vb_in_cnt = 0;
-  const uint32_t tag_len = 1;
-
-  memset(&ret_oid, 0, sizeof(ret_oid));
-  ret_oid.request = MIB_REQ_GET;
-
-  list_for_each_safe(curr, next, &sdg->vb_in_list) {
-    vb_in = list_entry(curr, struct var_bind, link);
-    vb_in_cnt++;
-
-    /* Search at the input oid */
-    mib_get(sdg, vb_in, &ret_oid);
-
-    val_len = ber_value_enc_try(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var));
-    vb_out = xmalloc(sizeof(*vb_out) + val_len);
-    vb_out->oid = ret_oid.oid;
-    vb_out->oid_len = ret_oid.id_len;
-    vb_out->value_type = tag(&ret_oid.var);
-    vb_out->value_len = ber_value_enc(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var), vb_out->value);
-
-    /* Error status */
-    if (ret_oid.err_stat) {
-      if (!sdg->pdu_hdr.err_stat) {
-        /* Mark the first error varbind */
-        sdg->pdu_hdr.err_stat = ret_oid.err_stat;
-        sdg->pdu_hdr.err_idx = vb_in_cnt;
-      }
-    }
-
-    /* OID length encoding */
-    oid_len = ber_value_enc_try(vb_out->oid, vb_out->oid_len, ASN1_TAG_OBJID);
-    len_len = ber_length_enc_try(oid_len);
-    vb_out->vb_len = tag_len + len_len + oid_len;
-
-    /* Value length encoding */
-    len_len = ber_length_enc_try(vb_out->value_len);
-    vb_out->vb_len += tag_len + len_len + vb_out->value_len;
-
-    /* Varbind length encoding */
-    len_len = ber_length_enc_try(vb_out->vb_len);
-    sdg->vb_list_len += tag_len + len_len + vb_out->vb_len;
-
-    /* Add into list. */
-    list_add_tail(&vb_out->link, &sdg->vb_out_list);
-    sdg->vb_out_cnt++;
-  }
-
-  snmp_response(sdg);
-}
-
-static void
-mib_getnext(struct snmp_datagram *sdg, struct var_bind *vb_in, struct oid_search_res *ret_oid)
-{
-  struct mib_view *view = NULL;
-  struct mib_community *community = NULL;
-  struct mib_user *user = NULL;
-
-  /* Access control */
-  if (sdg->version >= 3) {
-    user = mib_user_search(sdg->user_name);
-    if (user == NULL) {
-      ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-    }
-  } else {
-    community = mib_community_search(sdg->context_name);
-    if (community == NULL) {
-      ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-    }
-  }
-
-  /* Traverse all availble views according to community or user */
-  for (; ;) {
-    if (sdg->version >= 3) {
-        view = mib_user_next_view(user, MIB_ACES_READ, view);
-    } else {
-        view = mib_community_next_view(community, MIB_ACES_READ, view);
-    }
-
-    /* End of mib view */
-    if (view == NULL) {
-      /* Duplicate original oid when result not found */
-      ret_oid->oid = oid_dup(vb_in->oid, vb_in->oid_len);
-      ret_oid->id_len = vb_in->oid_len;
-      return;
-    }
-
-    mib_tree_search_next(view, vb_in->oid, vb_in->oid_len, ret_oid);
-    if (tag(&ret_oid->var) != ASN1_TAG_END_OF_MIB_VIEW) {
-      /* Gotcha */
-      break;
-    }
-    /* Free temporary result */
-    free(ret_oid->oid);
-  }
-}
-
-/* GETNEXT request function */
-static void
-snmp_getnext(struct snmp_datagram *sdg)
-{
-  struct list_head *curr, *next;
-  struct var_bind *vb_in, *vb_out;
-  struct oid_search_res ret_oid;
-  uint32_t oid_len, len_len, val_len;
-  uint32_t vb_in_cnt = 0;
-  const uint32_t tag_len = 1;
-
-  memset(&ret_oid, 0, sizeof(ret_oid));
-  ret_oid.request = MIB_REQ_GETNEXT;
-
-  list_for_each_safe(curr, next, &sdg->vb_in_list) {
-    vb_in = list_entry(curr, struct var_bind, link);
-    vb_in_cnt++;
-
-    /* Search at the next input oid */
-    mib_getnext(sdg, vb_in, &ret_oid);
-
-    val_len = ber_value_enc_try(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var));
-    vb_out = xmalloc(sizeof(*vb_out) + val_len);
-    vb_out->oid = ret_oid.oid;
-    vb_out->oid_len = ret_oid.id_len;
-    vb_out->value_type = tag(&ret_oid.var);
-    vb_out->value_len = ber_value_enc(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var), vb_out->value);
-
-    /* Error status */
-    if (ret_oid.err_stat) {
-      if (!sdg->pdu_hdr.err_stat) {
-        /* Report the first error varbind */
-        sdg->pdu_hdr.err_stat = ret_oid.err_stat;
-        sdg->pdu_hdr.err_idx = vb_in_cnt;
-      }
-    }
-
-    /* OID length encoding */
-    oid_len = ber_value_enc_try(vb_out->oid, vb_out->oid_len, ASN1_TAG_OBJID);
-    len_len = ber_length_enc_try(oid_len);
-    vb_out->vb_len = tag_len + len_len + oid_len;
-
-    /* Value length encoding */
-    len_len = ber_length_enc_try(vb_out->value_len);
-    vb_out->vb_len += tag_len + len_len + vb_out->value_len;
-
-    /* Varbind length encoding */
-    len_len = ber_length_enc_try(vb_out->vb_len);
-    sdg->vb_list_len += tag_len + len_len + vb_out->vb_len;
-
-    /* Add into list. */
-    list_add_tail(&vb_out->link, &sdg->vb_out_list);
-    sdg->vb_out_cnt++;
-  }
-
-  snmp_response(sdg);
-}
-
-static void
-mib_set(struct snmp_datagram *sdg, struct var_bind *vb_in, struct oid_search_res *ret_oid)
-{
-  struct mib_view *view = NULL;
-  struct mib_community *community = NULL;
-  struct mib_user *user = NULL;
-
-  /* Access control */
-  if (sdg->version >= 3) {
-    user = mib_user_search(sdg->user_name);
-    if (user != NULL) {
-      /* Check mib write views */
-      if (!mib_user_view_cover(user, MIB_ACES_WRITE, vb_in->oid, vb_in->oid_len)) {
-        ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-        user = NULL;
-      }
-    } else {
-      ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-    }
-  } else {
-    community = mib_community_search(sdg->context_name);
-    if (community != NULL) {
-      /* Check mib write views */
-      if (!mib_community_view_cover(community, MIB_ACES_WRITE, vb_in->oid, vb_in->oid_len)) {
-        ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-        community = NULL;
-      }
-    } else {
-      ret_oid->err_stat = SNMP_ERR_STAT_AUTHORIZATION;
-    }
-  }
-
-  /* Traverse all availble views according to community or user */
-  for (; ;) {
-    if (sdg->version >= 3) {
-        view = mib_user_next_view(user, MIB_ACES_READ, view);
-    } else {
-        view = mib_community_next_view(community, MIB_ACES_READ, view);
-    }
-
-    /* End of mib view */
-    if (view == NULL) {
-      /* Duplicate original oid when result not found */
-      ret_oid->oid = oid_dup(vb_in->oid, vb_in->oid_len);
-      ret_oid->id_len = vb_in->oid_len;
-      return;
-    }
-
-    mib_tree_search(view, vb_in->oid, vb_in->oid_len, ret_oid);
-    if ((!ret_oid->err_stat && MIB_TAG_VALID(tag(&ret_oid->var))) || oid_cmp(vb_in->oid, vb_in->oid_len, view->oid, view->id_len) < 0) {
-      /* Gotcha or given oid ahead of all views */
-      return;
-    }
-    /* Free temporary result */
-    free(ret_oid->oid);
-  }
-}
-
-/* SET request function */
-static void
-snmp_set(struct snmp_datagram *sdg)
-{
-  struct list_head *curr, *next;
-  struct var_bind *vb_in, *vb_out;
-  struct oid_search_res ret_oid;
-  uint32_t oid_len, len_len;
-  uint32_t vb_in_cnt = 0;
-  const uint32_t tag_len = 1;
-
-  memset(&ret_oid, 0, sizeof(ret_oid));
-  ret_oid.request = MIB_REQ_SET;
-
-  list_for_each_safe(curr, next, &sdg->vb_in_list) {
-    vb_in = list_entry(curr, struct var_bind, link);
-    vb_in_cnt++;
-
-    /* Decode the setting value ahead */
-    tag(&ret_oid.var) = vb_in->value_type;
-    length(&ret_oid.var) = ber_value_dec(vb_in->value, vb_in->value_len, tag(&ret_oid.var), value(&ret_oid.var));
-
-    /* Search at the input oid and set it */
-    mib_set(sdg, vb_in, &ret_oid);
-
-    vb_out = xmalloc(sizeof(*vb_out) + vb_in->value_len);
-    vb_out->oid = ret_oid.oid;
-    vb_out->oid_len = ret_oid.id_len;
-    vb_out->value_type = vb_in->value_type;
-    vb_out->value_len = ber_value_enc(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var), vb_out->value);
-
-    /* Invalid tags convert to error status for snmpset */
-    if (!ret_oid.err_stat && !MIB_TAG_VALID(tag(&ret_oid.var))) {
-      ret_oid.err_stat = SNMP_ERR_STAT_NOT_WRITABLE;
-    }
-
-    /* Error status */
-    if (ret_oid.err_stat) {
-      if (!sdg->pdu_hdr.err_stat) {
-        /* Report the first error varbind */
-        sdg->pdu_hdr.err_stat = ret_oid.err_stat;
-        sdg->pdu_hdr.err_idx = vb_in_cnt;
-      }
-    }
-
-    /* OID length encoding */
-    oid_len = ber_value_enc_try(vb_out->oid, vb_out->oid_len, ASN1_TAG_OBJID);
-    len_len = ber_length_enc_try(oid_len);
-    vb_out->vb_len = tag_len + len_len + oid_len;
-
-    /* Value length encoding */
-    len_len = ber_length_enc_try(vb_out->value_len);
-    vb_out->vb_len += tag_len + len_len + vb_out->value_len;
-
-    /* Varbind length encoding */
-    len_len = ber_length_enc_try(vb_out->vb_len);
-    sdg->vb_list_len += tag_len + len_len + vb_out->vb_len;
-
-    /* Add into list. */
-    list_add_tail(&vb_out->link, &sdg->vb_out_list);
-    sdg->vb_out_cnt++;
-  }
-
-  snmp_response(sdg);
-}
-
-/* BULKGET request function */
-void
-snmp_bulkget(struct snmp_datagram *sdg)
-{
-  struct list_head *curr, *next;
-  struct var_bind *vb_in, *vb_out;
-  struct oid_search_res ret_oid;
-  uint32_t oid_len, len_len, val_len;
-  uint32_t vb_in_cnt = 0;
-  uint32_t repeat;
-  const uint32_t tag_len = 1;
-
-  memset(&ret_oid, 0, sizeof(ret_oid));
-  ret_oid.request = MIB_REQ_GETNEXT;
-  repeat = sdg->pdu_hdr.err_idx;
-  sdg->pdu_hdr.err_idx = 0;
-
-  while (repeat-- > 0) {
-    list_for_each_safe(curr, next, &sdg->vb_in_list) {
-      vb_in = list_entry(curr, struct var_bind, link);
-      vb_in_cnt++;
-
-      /* Search at the next input oid */
-      mib_getnext(sdg, vb_in, &ret_oid);
-
-      /* Return oid for the next query. */
-      free(vb_in->oid);
-      vb_in->oid = oid_dup(ret_oid.oid, ret_oid.id_len);
-      vb_in->oid_len = ret_oid.id_len;
-
-      val_len = ber_value_enc_try(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var));
-      vb_out = xmalloc(sizeof(*vb_out) + val_len);
-      vb_out->oid = ret_oid.oid;
-      vb_out->oid_len = ret_oid.id_len;
-      vb_out->value_type = tag(&ret_oid.var);
-      vb_out->value_len = ber_value_enc(value(&ret_oid.var), length(&ret_oid.var), tag(&ret_oid.var), vb_out->value);
-
-      /* Error status */
-      if (ret_oid.err_stat) {
-        if (!sdg->pdu_hdr.err_stat) {
-          /* Report the first error varbind */
-          sdg->pdu_hdr.err_stat = ret_oid.err_stat;
-          sdg->pdu_hdr.err_idx = vb_in_cnt;
-        }
-      }
-
-      /* OID length encoding */
-      oid_len = ber_value_enc_try(vb_out->oid, vb_out->oid_len, ASN1_TAG_OBJID);
-      len_len = ber_length_enc_try(oid_len);
-      vb_out->vb_len = tag_len + len_len + oid_len;
-
-      /* Value length encoding */
-      len_len = ber_length_enc_try(vb_out->value_len);
-      vb_out->vb_len += tag_len + len_len + vb_out->value_len;
-
-      /* Varbind length encoding */
-      len_len = ber_length_enc_try(vb_out->vb_len);
-      sdg->vb_list_len += tag_len + len_len + vb_out->vb_len;
-
-      /* Add into list. */
-      list_add_tail(&vb_out->link, &sdg->vb_out_list);
-      sdg->vb_out_cnt++;
-    }
-  }
-
-  snmp_response(sdg);
-}
-
-/* Request dispatch */
-static void
-request_dispatch(struct snmp_datagram *sdg)
-{
-  switch (sdg->pdu_hdr.pdu_type) {
-    case MIB_REQ_GET:
-      if (sdg->vb_in_cnt == 0) {
-        sdg->pdu_hdr.pdu_type = MIB_REPO;
-      } else {
-        sdg->pdu_hdr.pdu_type = MIB_RESP;
-      }
-      snmp_get(sdg);
-      break;
-    case MIB_REQ_GETNEXT:
-      if (sdg->vb_in_cnt == 0) {
-        sdg->pdu_hdr.pdu_type = MIB_REPO;
-      } else {
-        sdg->pdu_hdr.pdu_type = MIB_RESP;
-      }
-      snmp_getnext(sdg);
-      break;
-    case MIB_RESP:
-      break;
-    case MIB_REQ_SET:
-      if (sdg->vb_in_cnt == 0) {
-        sdg->pdu_hdr.pdu_type = MIB_REPO;
-      } else {
-        sdg->pdu_hdr.pdu_type = MIB_RESP;
-      }
-      snmp_set(sdg);
-      break;
-    case MIB_REQ_BULKGET:
-      if (sdg->vb_in_cnt == 0) {
-        sdg->pdu_hdr.pdu_type = MIB_REPO;
-      } else {
-        sdg->pdu_hdr.pdu_type = MIB_RESP;
-      }
-      snmp_bulkget(sdg);
-      break;
-    case MIB_REQ_INF:
-      break;
-    case MIB_TRAP:
-      break;
-    case MIB_REPO:
-      break;
-    default:
-      break;
-  }
-}
-
 /* Alloc buffer for var bind decoding */
 static struct var_bind *
 var_bind_alloc(uint8_t *buf, enum snmp_err_code *err)
@@ -970,6 +514,56 @@ DECODE_FINISH:
   }
 }
 
+/* SNMP request dispatch */
+static void
+snmp_request_dispatch(struct snmp_datagram *sdg)
+{
+  switch (sdg->pdu_hdr.pdu_type) {
+    case MIB_REQ_GET:
+      if (sdg->vb_in_cnt == 0) {
+        sdg->pdu_hdr.pdu_type = MIB_REPO;
+      } else {
+        sdg->pdu_hdr.pdu_type = MIB_RESP;
+      }
+      snmp_get(sdg);
+      break;
+    case MIB_REQ_GETNEXT:
+      if (sdg->vb_in_cnt == 0) {
+        sdg->pdu_hdr.pdu_type = MIB_REPO;
+      } else {
+        sdg->pdu_hdr.pdu_type = MIB_RESP;
+      }
+      snmp_getnext(sdg);
+      break;
+    case MIB_RESP:
+      break;
+    case MIB_REQ_SET:
+      if (sdg->vb_in_cnt == 0) {
+        sdg->pdu_hdr.pdu_type = MIB_REPO;
+      } else {
+        sdg->pdu_hdr.pdu_type = MIB_RESP;
+      }
+      snmp_set(sdg);
+      break;
+    case MIB_REQ_BULKGET:
+      if (sdg->vb_in_cnt == 0) {
+        sdg->pdu_hdr.pdu_type = MIB_REPO;
+      } else {
+        sdg->pdu_hdr.pdu_type = MIB_RESP;
+      }
+      snmp_bulkget(sdg);
+      break;
+    case MIB_REQ_INF:
+      break;
+    case MIB_TRAP:
+      break;
+    case MIB_REPO:
+      break;
+    default:
+      break;
+  }
+}
+
 /* Receive snmp datagram from transport module */
 void
 snmpd_recv(uint8_t *buffer, int len)
@@ -1002,5 +596,5 @@ snmpd_recv(uint8_t *buffer, int len)
   snmp_decode(&snmp_datagram);
 
   /* Dispatch request */
-  request_dispatch(&snmp_datagram);
+  snmp_request_dispatch(&snmp_datagram);
 }
